server/requirements.txt
Flask
Flask-Cors
PyHamcrest
